feat: refactor packet forward middleware and more
